default image logic ok

+++ app/controllers/ImageController.php
@@ -33,6 +33,12 @@ class ImageController extends Controller {
 		$img->news_id = $img_id;
 
 		$img->save();
+
+		DB::table('images')
+			->where('name', '=', 'default.gif')
+			->where('news_id', '=', $img_id)
+			->delete();
+
 		return $img_name;
 	}
 
@@ -47,20 +53,34 @@ class ImageController extends Controller {
 	//selected album images delete 
 	public function postDestroy(){
 		$images = Input::get('imgStatus');
+
 		if($images){
 			for ($i=0; $i < sizeof($images); $i++) { 
 				$image = Image::find($images[$i]);
-
-				$image->delete();
-				$path = "uploads/images/".$image->name;
-				
-				if(file_exists($path)){
-					unlink($path);
-				}
+				if($image->name != 'default.gif'){
+					$image->delete();
+					$path = "uploads/images/".$image->name;
+					
+					if(file_exists($path)){
+						unlink($path);
+					}		
+	
+			}else{
+				return Redirect::to('admin/news/index')
+					->with('message', 'Add images to remove default image for the news');
 			}
+			
+		}
+		if(!DB::table('images')->where('news_id', '=', $image->news_id)->first()){
+			$def = new Image;
+			$def->name = 'default.gif';
+			$def->news_id = $image->news_id;
+			$def->save();
+		}
 
-			return Redirect::to('admin/news/index')
-			->with('message', 'Images deleted successfully');
+		return Redirect::to('admin/news/index')
+					->with('message', 'Images deleted successfully');			
+				
 		}
 
 		return Redirect::to('admin/news/index')
